Finish simplification of `Gpu` access
