+++ crates/modor_graphics/src/resources.rs
@@ -1,6 +1,7 @@
 use crate::mesh::Mesh;
-use crate::{Camera2D, DefaultMaterial2D, Mat, Window};
+use crate::{Camera2D, Mat, Window};
 use modor::{Context, Node, RootNode, Visit};
+use crate::material::default_2d::DefaultMaterial2D;
 
 #[non_exhaustive]
 #[derive(Debug, Node, Visit)]
