+++ crates/modor_resources/src/resource.rs
@@ -154,7 +154,7 @@ where
     }
 }
 
-// TODO: add loaded_version() method to easilyl identify when the new loading is done
+// TODO: add loaded_version() method to easily identify when the new loading is done
 impl<T> Res<T>
 where
     T: Resource,
